+++ ios/Classes/SwiftHealthPlugin.swift
@@ -234,6 +234,8 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         unitDict[HealthConstants.RESPIRATIONS_PER_MINUTE] = HKUnit(from: "count/min")
         unitDict[HealthConstants.MILLIGRAM_PER_DECILITER] = HKUnit(from: "mg/dL")
         unitDict[HealthConstants.METER_PER_SECOND] = HKUnit(from: "m/s")
+        unitDict[HealthConstants.WATT] = HKUnit(from: "W")
+        unitDict[HealthConstants.REVOLUTIONS_PER_MINUTE] = HKUnit(from: "count/min")
         unitDict[HealthConstants.UNKNOWN_UNIT] = HKUnit(from: "")
         unitDict[HealthConstants.NO_UNIT] = HKUnit(from: "")
 
@@ -310,6 +312,14 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
 
         if #available(iOS 16.0, *) {
             initializeIOS16Types()
+        }
+
+        if #available(iOS 17.0, *) {
+            initializeIOS17Types()
+        }
+
+        if #available(iOS 18.0, *) {
+            initializeIOS18Types()
         }
 
         // Concatenate heart events, headache and health data types (both may be empty)
@@ -551,9 +561,35 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
         dataTypesDict[HealthConstants.UNDERWATER_DEPTH] = HKQuantityType.quantityType(forIdentifier: .underwaterDepth)!
         dataTypesDict[HealthConstants.UV_INDEX] = HKQuantityType.quantityType(forIdentifier: .uvExposure)!
         dataTypesDict[HealthConstants.SLEEP_WRIST_TEMPERATURE] = HKQuantityType.quantityType(forIdentifier: .appleSleepingWristTemperature)!
+        dataTypesDict[HealthConstants.RUNNING_SPEED] = HKQuantityType.quantityType(forIdentifier: .runningSpeed)!
+        dataTypesDict[HealthConstants.RUNNING_POWER] = HKQuantityType.quantityType(forIdentifier: .runningPower)!
 
         dataQuantityTypesDict[HealthConstants.UV_INDEX] = HKQuantityType.quantityType(forIdentifier: .uvExposure)!
         dataQuantityTypesDict[HealthConstants.SLEEP_WRIST_TEMPERATURE] = HKQuantityType.quantityType(forIdentifier: .appleSleepingWristTemperature)!
+        dataQuantityTypesDict[HealthConstants.RUNNING_SPEED] = HKQuantityType.quantityType(forIdentifier: .runningSpeed)!
+        dataQuantityTypesDict[HealthConstants.RUNNING_POWER] = HKQuantityType.quantityType(forIdentifier: .runningPower)!
+    }
+
+    /// Initialize iOS 17 specific data types.
+    @available(iOS 17.0, *)
+    private func initializeIOS17Types() {
+        dataTypesDict[HealthConstants.CYCLING_SPEED] = HKQuantityType.quantityType(forIdentifier: .cyclingSpeed)!
+        dataTypesDict[HealthConstants.CYCLING_POWER] = HKQuantityType.quantityType(forIdentifier: .cyclingPower)!
+        dataTypesDict[HealthConstants.CYCLING_CADENCE] = HKQuantityType.quantityType(forIdentifier: .cyclingCadence)!
+
+        dataQuantityTypesDict[HealthConstants.CYCLING_SPEED] = HKQuantityType.quantityType(forIdentifier: .cyclingSpeed)!
+        dataQuantityTypesDict[HealthConstants.CYCLING_POWER] = HKQuantityType.quantityType(forIdentifier: .cyclingPower)!
+        dataQuantityTypesDict[HealthConstants.CYCLING_CADENCE] = HKQuantityType.quantityType(forIdentifier: .cyclingCadence)!
+    }
+
+    /// Initialize iOS 18 specific data types.
+    @available(iOS 18.0, *)
+    private func initializeIOS18Types() {
+        dataTypesDict[HealthConstants.ROWING_SPEED] = HKQuantityType.quantityType(forIdentifier: .rowingSpeed)!
+        dataTypesDict[HealthConstants.DISTANCE_ROWING] = HKQuantityType.quantityType(forIdentifier: .distanceRowing)!
+
+        dataQuantityTypesDict[HealthConstants.ROWING_SPEED] = HKQuantityType.quantityType(forIdentifier: .rowingSpeed)!
+        dataQuantityTypesDict[HealthConstants.DISTANCE_ROWING] = HKQuantityType.quantityType(forIdentifier: .distanceRowing)!
     }
 
     /// Initialize workout activity types